+++ icevision/core/record_components.py
@@ -278,8 +278,24 @@ class InstancesLabelsRecordComponent(BaseLabelsRecordComponent):
 
 
 class ClassificationLabelsRecordComponent(BaseLabelsRecordComponent):
-    def __init__(self, task=tasks.classification):
+    def __init__(self, task=tasks.classification, is_multilabel: bool = False):
         super().__init__(task=task)
+        self.is_multilabel = is_multilabel
+
+    def _autofix(self):
+        if not self.is_multilabel and len(self.labels) > 1:
+            raise AutofixAbort(
+                f"Expected a single label, got {len(self.labels)} instead. "
+                f"If you want to do multi-label classification, initiate the record "
+                f"with `is_multilabel=True`"
+            )
+        return super()._autofix()
+
+    def one_hot_encoded(self) -> np.array:
+        "Get labels as a one-hot encoded array"
+        one_hot_labels = np.zeros(len(self.class_map))
+        one_hot_labels[self.label_ids] = 1
+        return one_hot_labels
 
 
 class BBoxesRecordComponent(RecordComponent):
@@ -298,11 +314,15 @@ class BBoxesRecordComponent(RecordComponent):
         for bbox in self.bboxes:
             try:
                 autofixed = bbox.autofix(
-                    img_w=self.composite.width, img_h=self.composite.height
+                    img_w=self.composite.width,
+                    img_h=self.composite.height,
+                    record_id=self.composite.record_id,
                 )
                 success.append(True)
             except InvalidDataError as e:
-                logger.log("AUTOFIX-FAIL", "{}", str(e))
+                autofix_log(
+                    "AUTOFIX-FAIL", "{}", str(e), record_id=self.composite.record_id
+                )
                 success.append(False)
 
         return {"bboxes": success}