+++ icevision/metrics/confusion_matrix/__init__.py
@@ -1,0 +1,1 @@
+from icevision.metrics.confusion_matrix.confusion_matrix import *